requirements.txt
celery==5.3.6
fastapi[all]==0.110.0
httpx==0.27.0
hubspot-api-client==8.2.1
pip==24.0
pyairtable==2.2.2
pydantic==2.3.0
pydantic-core==2.6.3
pydantic-settings==2.2.1
PyJWT==2.8.0
python-dotenv==1.0.0
PyYAML==6.0.1
redis==5.0.1
requests==2.31.0
ShopifyAPI==12.3.0
starlette==0.36.3
typing-extensions==4.9.0
uvicorn[standard]==0.20.0
gunicorn==21.2.0
websockets==12.0